+++ src/methods/diff/index.js
@@ -1,8 +1,8 @@
 const fns = require('../../fns')
 const waterfall = require('./waterfall')
 
-const reverseDiff = function(obj) {
-  Object.keys(obj).forEach(k => {
+const reverseDiff = function (obj) {
+  Object.keys(obj).forEach((k) => {
     obj[k] *= -1
   })
   return obj
@@ -11,7 +11,7 @@ const reverseDiff = function(obj) {
 // this method counts a total # of each unit, between a, b.
 // '1 month' means 28 days in february
 // '1 year' means 366 days in a leap year
-const main = function(a, b, unit) {
+const main = function (a, b, unit) {
   b = fns.beADate(b, a)
   //reverse values, if necessary
   let reversed = false